requirements.txt
faiss-cpu
numpy
sentence-transformers
PyMuPDF
requests
python-dotenv
mcp